More date time fixes
