+++ orchestra/application/utility/generalhelper.cpp
@@ -25,6 +25,7 @@
 #include <mutex>
 #include <random>
 #include <string>
+#include <datetime/datetime.hpp>
 #include <general.hpp>  // pscore utility
 
 namespace app {
